requirements.txt
fastapi[all]
sqlalchemy[asyncio]
asyncpg
uvicorn
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
email-validator
python-dotenv
bcrypt==3.2.0
passlib==1.7.4

# # requirements.txt
# fastapi==0.115.0
# uvicorn[standard]==0.30.1
# sqlalchemy[asyncio]==2.0.30
# asyncpg==0.29.0
# alembic==1.13.2
# python-dotenv==1.0.1
# pydantic==2.7.4
# pydantic-settings==2.2.1
# passlib[bcrypt]==1.7.4
# python-jose[cryptography]==3.3.0
